--- src/lib/kotlin/slatekit-policy/src/main/kotlin/slatekit/policy/Input.kt
@@ -1,7 +1,0 @@
-package slatekit.policy
-
-import slatekit.results.Outcome
-
-interface Input<I> {
-    suspend fun process(i: Outcome<I>): Outcome<I>
-}
