--- src/lib/kotlin/slatekit-policy/src/main/kotlin/slatekit/policy/Output.kt
@@ -1,7 +1,0 @@
-package slatekit.policy
-
-import slatekit.results.Outcome
-
-interface Output<I, O> {
-    suspend fun process(raw:I, i: Outcome<I>, o: Outcome<O>): Outcome<O>
-}
