--- src/lib/kotlin/slatekit-policy/src/main/kotlin/slatekit/policy/Process.kt
@@ -1,7 +1,0 @@
-package slatekit.policy
-
-import slatekit.results.Outcome
-
-interface Process<I, O> {
-    suspend fun process(i: I, next: suspend(I) -> Outcome<O>): Outcome<O>
-}
