+++ src/lib/kotlin/slatekit-policy/src/main/kotlin/slatekit/policy/hooks/After.kt
@@ -1,17 +1,21 @@
-package slatekit.policy.middleware
+package slatekit.policy.hooks
 
 import slatekit.common.Ignore
+import slatekit.results.Outcome
 
 /**
  * A "Hooks" based middle-ware that allows only handling before/after events
  * of a call, without any modification to the life-cycle/flow.
  */
-interface Before<TReq> : Middleware {
-
+interface After<TReq, TRes> {
     /**
      * Middleware hook for after a request is made
      * @param req : The request for the call
+     * @param res : The result of the call
      */
     @Ignore
-    suspend fun before(req: TReq)
+    suspend fun after(req:TReq, res: Outcome<TRes>)
 }
+
+
+
