+++ src/lib/kotlin/slatekit-policy/src/main/kotlin/slatekit/policy/hooks/Before.kt
@@ -1,21 +1,17 @@
-package slatekit.policy.middleware
+package slatekit.policy.hooks
 
 import slatekit.common.Ignore
-import slatekit.results.Outcome
 
 /**
  * A "Hooks" based middle-ware that allows only handling before/after events
  * of a call, without any modification to the life-cycle/flow.
  */
-interface After<TReq, TRes> {
+interface Before<TReq> : Middleware {
+
     /**
      * Middleware hook for after a request is made
      * @param req : The request for the call
-     * @param res : The result of the call
      */
     @Ignore
-    suspend fun after(raw:TReq, req: Outcome<TReq>, res: Outcome<TRes>)
+    suspend fun before(req: TReq)
 }
-
-
-
